Putting definitions of SCD and SCDMatcher separated from sc_dis.cpp file
